RL_researcher_writer_ymaxing/agents_integration_local/mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Multi-server MCP client for Nova and Brown"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.9.0",
    "pydantic-settings>=2.10.0",
    "mcp>=1.11.0",
    "google-genai>=1.20.0",
    "opik>=1.8.42",
]
